+++ config/medium/Fibonomials/code/CGCode.js
@@ -1,0 +1,12 @@
+const definedBy = readline();
+const aI = readline();
+const bI = readline();
+const x = parseInt(readline());
+const n = parseInt(readline());
+for (let i = 0; i < n; i++) {
+
+    // Write an answer using console.log()
+    // To debug: console.error('Debug messages...');
+
+    console.log('p_i(x)');
+}